+++ 2-Computer-Systems/problems/1-Protobuf-Varint/src/Mathematical/Utilities/BinaryEncoding/getNumberBitSequence.ts
