+++ 2-Computer-Systems/problems/1-Protobuf-Varint/src/Mathematical/Utilities/HexEncoding/hexTypes.d.ts
